+++ Chimp/Views/NFCOperationCoordinator.swift
@@ -10,15 +10,15 @@ class NFCOperationCoordinator: NSObject {
     private let nfcManager = NFCManager()
     
     // Callbacks (preserved for HomeViewModel integration)
-    var onLoadNFTSuccess: ((String, UInt64) -> Void)?
+    var onLoadNFTSuccess: ((String, UInt32) -> Void)?
     var onLoadNFTError: ((String) -> Void)?
-    var onClaimSuccess: ((UInt64, String) -> Void)? // tokenId, contractId
+    var onClaimSuccess: ((UInt32, String) -> Void)? // tokenId, contractId
     var onClaimError: ((String) -> Void)?
     var onTransferSuccess: (() -> Void)?
     var onTransferError: ((String) -> Void)?
     var onSignSuccess: ((UInt32, UInt32, String) -> Void)? // globalCounter, keyCounter, signature
     var onSignError: ((String) -> Void)?
-    var onMintSuccess: ((UInt64) -> Void)? // tokenId
+    var onMintSuccess: ((UInt32) -> Void)? // tokenId
     var onMintError: ((String) -> Void)?
     
     // MARK: - Load NFT
@@ -46,7 +46,7 @@ class NFCOperationCoordinator: NSObject {
     }
     
     // MARK: - Read NFT for Transfer (first scan to get token ID)
-    func readNFTForTransfer(completion: @escaping (Bool, UInt64?, String?) -> Void) {
+    func readNFTForTransfer(completion: @escaping (Bool, UInt32?, String?) -> Void) {
         nfcManager.readNFTForTransfer { success, tokenId, error in
             completion(success, tokenId, error)
             // No specific callbacks for this operation - just direct completion
@@ -54,7 +54,7 @@ class NFCOperationCoordinator: NSObject {
     }
     
     // MARK: - Transfer NFT (second scan to complete transfer)
-    func transferNFT(recipientAddress: String, tokenId: UInt64, completion: @escaping (Bool, String?) -> Void) {
+    func transferNFT(recipientAddress: String, tokenId: UInt32, completion: @escaping (Bool, String?) -> Void) {
         nfcManager.transferNFT(recipientAddress: recipientAddress, tokenId: tokenId) { [weak self] success, error in
             completion(success, error)
             if success {
